backend/requirements-offline.txt
crcmod==1.7
fastapi-offline==1.7.5
matplotlib==3.10.3
networkx==3.4.2
numpy==2.2.6
pydantic==2.11.7
python-snap7==2.0.2
SQLAlchemy==2.0.41
uvicorn==0.37.0
pyinstaller
py7zr